Added code to find price of Apts
SO.. were back here again. This time we did a LOT of work in the span of about 2 hours?? or so. something like that. So first it was like 10 lines of code (something like that), but then i put on my cool shades and shortened it down to a WHOPPING 4 lines of code, which if you take out the comments and the last print statement, its like all on one line of code.

Thats right folks.

I am a coding now. :smirk: :cool:

*2 hours of work for one line. such O fish Ant. Very wow.