Added availability finding and output list
Okay, we're back after a short hiatus. Thanks for stickin with me :)

This time we put our big brains together to make an output list variable so that we could add up all the data we are collecting into one list per iteration ( each iteration would be a different apartment  )

The main stuff in this commit though is finding availability.

Progress is slow, but progress is progress.     -Mark Zuckerberg

Did he really say that?       -Chris

I'll leave that for you to decide.